+++ app/src/container/MapContainer/index.js
@@ -3,8 +3,8 @@ import * as React from 'react';
 import {Alert, Platform} from 'react-native';
 import {connect} from 'react-redux';
 import Map from '../../screens/Map';
-import {fetchParkspots, filterParkspots, updateLocation, updateMapPosition} from './actions';
-import {clearSelectedLocation} from '../SearchContainer/actions';
+import {addFavoriteByDescription, fetchParkspots, filterParkspots, updateLocation, updateMapPosition} from './actions';
+import {clearSelectedLocation, remFavorite} from '../SearchContainer/actions';
 
 export interface Props {
   navigation: any;
@@ -17,6 +17,9 @@ export interface Props {
   selectedLocation: Object;
   filterParkspots: Function;
   clearSelectedLocation: Function;
+  addFavoriteByDescription: Function;
+  remFavorite: Function;
+  favorites: Object;
 }
 
 export interface State {
@@ -36,6 +39,9 @@ class MapContainer extends React.Component<Props, State> {
         selectedLocation={this.props.selectedLocation}
         filterParkspots={this.props.filterParkspots}
         clearSelectedLocation={this.props.clearSelectedLocation}
+        addFavoriteByDescription={this.props.addFavoriteByDescription}
+        remFavorite={this.props.remFavorite}
+        favorites={this.props.favorites}
       />
     );
   }
@@ -54,6 +60,8 @@ function bindAction(dispatch) {
     updateMapPosition: (mapPosition) => dispatch(updateMapPosition(mapPosition)),
     filterParkspots: (filterId) => dispatch(filterParkspots(filterId)),
     clearSelectedLocation: () => dispatch(clearSelectedLocation()),
+    addFavoriteByDescription: (description) => dispatch(addFavoriteByDescription(description)),
+    remFavorite: (fav) => dispatch(remFavorite(fav)),
   };
 }
 
@@ -61,6 +69,7 @@ const mapStateToProps = state => ({
   parkspots: state.mapReducer.parkspots,
   userPosition: state.mapReducer.userPosition,
   mapPosition: state.mapReducer.mapPosition,
-  selectedLocation: state.searchReducer.selectedLocation
+  selectedLocation: state.searchReducer.selectedLocation,
+  favorites: state.searchReducer.favorites,
 });
 export default connect(mapStateToProps, bindAction)(MapContainer);