+++ app/src/screens/Map/index.js
@@ -46,6 +46,9 @@ export interface Props {
   selectedLocation: Object;
   filterParkspots: Function;
   clearSelectedLocation: Function;
+  addFavoriteByDescription: Function;
+  remFavorite: Function;
+  favorites: Object;
 }
 
 export interface State {
@@ -90,10 +93,23 @@ class Map extends React.Component<Props, State> {
     this.setShowFilters(false);
     this.setState({
       selectedParkspot: this.props.parkspots.find(parkspot => {
-        return (
+        if (
           Number(parkspot.lat) === Number(event.nativeEvent.coordinate.latitude) &&
           Number(parkspot.lng) === Number(event.nativeEvent.coordinate.longitude)
-        );
+        ) {
+          // Check if location is added to Favorites
+          parkspot.favorite = false;
+          parkspot.locationId = null;
+          parkspot.description = parkspot.street+' '+parkspot.houseNumber+', '+parkspot.city+', '+parkspot.country;
+          this.props.favorites.find((fav) => {
+            if (fav.description === parkspot.description) {
+              parkspot.favorite = true;
+              parkspot.locationId = fav.id;
+              return true;
+            }
+          });
+          return true;
+        }
       }),
     });
   };
@@ -423,6 +439,30 @@ class Map extends React.Component<Props, State> {
     this.map = undefined;
   }
 
+  componentWillReceiveProps(nextProps) {
+    // If favorite field changed from the mapCard we need to update the selectedParkspot
+    if (this.state.selectedParkspot && this.props.favorites.length != nextProps.favorites.length) {
+      let parkspot = this.state.selectedParkspot;
+      const found = nextProps.favorites.find((fav) => {
+        if (fav.description === parkspot.description) {
+          parkspot.favorite = true;
+          parkspot.locationId = fav.id;
+          this.setState({
+            selectedParkspot: parkspot
+          })
+          return true;
+        }
+      });
+      if (!found){
+        parkspot.favorite = false;
+        parkspot.locationId = null;
+        this.setState({
+          selectedParkspot: parkspot
+        })
+      }
+    }
+  }
+
   componentDidMount() {
     //remove after DEV
     if (!__DEV__) {
@@ -497,6 +537,8 @@ class Map extends React.Component<Props, State> {
           drivingDirections={this.state.drivingDirections}
           walkingDirections={this.state.walkingDirections}
           destinationName={this.props.selectedLocation ? this.props.selectedLocation.description : null}
+          addFavoriteByDescription={this.props.addFavoriteByDescription}
+          remFavorite={this.props.remFavorite}
         />
         }